+++ src/components/datasets-grid/constants/index.ts
@@ -1,11 +1,9 @@
-export type Theme =
-  | 'Water'
-  | 'Agriculture'
-  | 'Biodiversity'
-  | 'Soil'
-  | 'Climate & Health'
-  | 'Forest';
+import type { SortingCriteria, Theme } from '../types';
 
+export const SORTING = [
+  'title',
+  //  'date' TO - Do add sorting by date when API gets ready
+] satisfies SortingCriteria[];
 export const THEMES = [
   'Water',
   'Agriculture',