+++ src/components/ui/radio-group.tsx
@@ -9,14 +9,19 @@ import { cn } from '@/lib/classnames';
 
 const RadioGroup = React.forwardRef<
   React.ElementRef<typeof RadioGroupPrimitive.Root>,
-  React.ComponentPropsWithoutRef<typeof RadioGroupPrimitive.Root>
->(({ className, ...props }, ref) => {
+  React.ComponentPropsWithoutRef<typeof RadioGroupPrimitive.Root> & {
+    children: React.ReactNode;
+    className?: string;
+  }
+>(({ className, children, ...props }, ref) => {
   return (
     <RadioGroupPrimitive.Root
-      className={cn({ 'grid gap-2': true, [className]: !!className })}
+      className={cn({ 'grid gap-2 text-secondary-500': true, [className]: !!className })}
       {...props}
       ref={ref}
-    />
+    >
+      {children}
+    </RadioGroupPrimitive.Root>
   );
 });
 RadioGroup.displayName = RadioGroupPrimitive.Root.displayName;
@@ -28,13 +33,14 @@ const RadioGroupItem = React.forwardRef<
   return (
     <RadioGroupPrimitive.Item
       ref={ref}
-      className={cn(
-        'h-5 w-5 rounded-full border-2 border-primary border-secondary-500 text-primary ring-offset-background focus:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50',
-        className
-      )}
+      className={cn({
+        'h-5 w-5 rounded-full border-2 border-primary border-secondary-500 ring-offset-background focus:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50':
+          true,
+        [className]: !!className,
+      })}
       {...props}
     >
-      <RadioGroupPrimitive.Indicator className="flex items-center justify-center rounded-full text-secondary-500">
+      <RadioGroupPrimitive.Indicator className="flex items-center justify-center rounded-full">
         <Circle className="h-2.5 w-2.5 fill-current text-current" />
       </RadioGroupPrimitive.Indicator>
     </RadioGroupPrimitive.Item>